# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "CIMOL API with Golang.\nGolang memiliki keunggulan dalam skalabilitas horizontal dan pemeliharaan kode yang lebih sederhana,\nGolang juga memiliki keamanan bawaan dan memudahkan pencegahan beberapa jenis serangan, serta\nkinerja yang signifikan dalam kasus pengolahan beban kerja tinggi dan pembuatan API yang membutuhkan respons cepat.",
        "title": "CIMOL API with Golang",
        "contact": {
            "name": "API CNAF MOBILE",
            "url": "https://www.cnaf.co.id/",
            "email": "dimas.zikri@cnaf.co.id"
        },
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/api",
    "paths": {
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "AUTH"
                ],
                "summary": "Login into API Cimol",
                "parameters": [
                    {
                        "description": "Payload Body [RAW]",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserLogin"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseWithoutData"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "AUTH"
                ],
                "summary": "Logout API Cimol",
                "parameters": [
                    {
                        "description": "Payload Body [RAW]",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Login"
                        }
                    },
                    {
                        "type": "string",
                        "description": "string example",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseWithoutData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseWithoutData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "helpers.ResponseWithoutData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "responsecode": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.Login": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "pass,123"
                },
                "username": {
                    "type": "string",
                    "example": "1879"
                }
            }
        },
...
        "models.UserLogin": {
            "type": "object",
            "properties": {
                "base64qrcode": {
                    "type": "string"
                },
                "bucketcoll": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "greetings": {
                    "type": "string"
                },
                "hp": {
                    "type": "string"
                },
                "last_login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "realname": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        }
    }
}