# helpers/response.go
package helpers

import (
	"encoding/json"
	"net/http"
)

type ResponseWithData struct {
	ResponseCode int    `json:"responsecode"`
	Status       string `json:"status"`
	Message      string `json:"message"`
	Data         any    `json:"data"`
}

type ResponseWithoutData struct {
	ResponseCode int    `json:"responsecode"`
	Status       string `json:"status"`
	Message      string `json:"message"`
}

func Response(w http.ResponseWriter, code int, message string, payload interface{}) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(code)

	var response any
	status := "success"

	if code >= 400 {
		status = "failed"
	}

	if payload != nil {
		response = &ResponseWithData{
			ResponseCode: code,
			Status:       status,
			Message:      message,
			Data:         payload,
		}
	} else {
		response = &ResponseWithoutData{
			ResponseCode: code,
			Status:       status,
			Message:      message,
		}
	}

	res, _ := json.Marshal(response)
	w.Write(res)
}